+++ src/components/footerN/footer.component.tsx
