+++ src/components/footerN/footer.styles.ts
