+++ src/components/footerN/index.ts
