+++ src/components/loaderN/index.ts
