+++ src/components/loaderN/loader.component.tsx
