+++ src/components/loaderN/loader.styles.ts
