+++ src/pages/paintN/components/buttons-panel/buttons-panel.component.tsx
