+++ src/pages/paintN/components/buttons-panel/buttons-panel.styles.ts
