+++ src/pages/paintN/components/buttons-panel/index.ts
