+++ src/pages/paintN/components/drawing-tools-panel/drawing-tools-panel.styles.ts
