+++ src/pages/paintN/index.ts
