+++ src/pages/paintN/paint.component.tsx
