+++ src/pages/paintN/paint.styles.ts
